% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.r
\name{data_info}
\alias{data_info}
\title{Get an info about the dataset in vimo}
\usage{
data_info(dataset_id)
}
\arguments{
\item{model_name}{Name of the model in **vimo**, character}
}
\value{
named list containing all meta data about model
}
\description{
This tool is used for getting a meta data about the model that is already uploaded to **vimo**.
}
\examples{
library("vimo")

model_info("example_model")
model_info("example_model")$model_info
model_info("example_model")$data_info
model_info("example_model")$data_info$dataset_id

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
