% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_data}
\alias{get_data}
\title{Get the dataset from **vimo**}
\usage{
get_data(dataset_id)
}
\arguments{
\item{dataset_id}{hash of the dataset}
}
\value{
data frame
}
\description{
You can use this function to download the dataset from vimo as a data frame
}
\examples{
library("vimo")

get_data('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
