% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_data.r
\name{head_data}
\alias{head_data}
\title{Get the head of the dataset in the vimo}
\usage{
head_data(dataset_id, n = 5)
}
\arguments{
\item{dataset_id}{the dataset hash}

\item{n}{number of rows to show}
}
\value{
top n rows of the dataset
}
\description{
This tool allows you to view the head of the dataset in the vimo.
}
\examples{
library("vimo")

head_data('aaaaaaaaaaaaaaaaaaaaaaaaaaaaa')

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
