% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_audit.r
\name{model_audit}
\alias{model_audit}
\title{Make an audit of the model in the vimo}
\usage{
model_audit(model_name, measure, user, password, data, target,
  data_name = NA, data_desc = NA)
}
\arguments{
\item{model_name}{name of the model in the vimo, character}

\item{measure}{name of the measure used in the audit, character}

\item{user}{your user name, character}

\item{password}{your password}

\item{data}{data frame to make an audit on or path or hash of already uploaded dataset}

\item{target}{name of the target column in the dataset}

\item{data_name}{name of the dataset that will be visible in the vimo, unnecessary if data is a hash}

\item{data_desc}{description of the dataset, unnecessary if data is a hash}
}
\value{
result of the audit or information if somethin went wrong
}
\description{
You can use this function to audit in different ways models already uploaded in the vimo.
}
\examples{
\\code{
library("vimo")

audit_model('example_model', 'mae', 'Example user', 'example password', iris, 'Species', 'iris', 'Flowers')
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
