% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_upload.r
\name{model_upload}
\alias{model_upload}
\title{Upload the model to vimo}
\usage{
model_upload(model, model_name, model_desc, target, tags, train_dataset,
  train_dataset_name, dataset_desc, user_name, password)
}
\arguments{
\item{model}{mlr, caret or parsnip model, or path to it written as '.RDS' file}

\item{model_name}{name of the model that will be visible in the vimo}

\item{model_desc}{description of the model}

\item{tags}{vector of model's tags, should be vector of strings}

\item{train_dataset}{training dataset with named columns and target column as the last one, or path to *.csv* file (must contain **/** sign) or hash of already
uploaded data}

\item{train_dataset_name}{name of the training dataset that will be visible in the database}

\item{dataset_desc}{description of the dataset}

\item{user_name}{your user name}

\item{password}{your password}
}
\description{
This tool allows you to upload your mlr, caret or parsnip model to model governance base vimo.
}
\examples{
\\code{
library("vimo")
library("mlr")

task = makeClassifTask(data = iris, target = 'Species')
model = makeLearner("classif.randomForest")
model = train(model, task)

upload(model, "example_model", "This is an example model", c('example', 'easy'), iris, "example_training_data", 'This is an example data', 'Example user', 
'example password')
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
