% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.character}
\alias{predict.character}
\title{Make a prediction using **vimo** model}
\usage{
\method{predict}{character}(model_name, X, pred_type = "exact",
  prepare_columns = TRUE)
}
\arguments{
\item{model_name}{name of the model in **vimo**}

\item{pred_type}{type of prediction, 'exact' or 'prob'}

\item{prepare_columns}{if X is an object then columns' names will be fetched automatically}
}
\description{
This tool allows you to make a prediction with model in **vimo**.
}
\examples{
\\code {
library("vimo")

predict("example_model", iris[,-5])
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
