% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict}
\alias{predict}
\title{Make a prediction using **vimo** model}
\usage{
predict(model_name, X, pred_type = "exact", prepare_columns = TRUE)
}
\arguments{
\item{model_name}{name of the model in **vimo**}
}
\description{
This tool allows you to make a prediction of model in **vimo**.
}
\examples{
library("vimo")

predict("example_model", iris[,-5])

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
