% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_model.r
\name{search_model}
\alias{search_model}
\title{Get an info about the model in **vimo**}
\usage{
search_model(tags)
}
\arguments{
\item{model_name}{Name of the model in **vimo**, character}
}
\value{
named list containing all meta data about model
}
\description{
This tool is used for getting a meta data about the model that is already uploaded to **vimo**.
}
\examples{
library("vimo")

model_info("example_model")

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
