% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.r
\name{upload_data}
\alias{upload_data}
\title{Upload the dataset to the vimo}
\usage{
upload_data(data, data_name, user_name, password)
}
\arguments{
\item{data}{the data frame to upload or path}

\item{data_name}{name of the dataset that will be visible in the vimo}

\item{user_name}{your user name}

\item{password}{your password}
}
\value{
information if uploading the data was successful
}
\description{
This function uploads the dataset and needed metadata to the vimo.
}
\examples{
library("vimo")

upload_data(X, 'data name', 'Example user', 'example password')

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
