% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_create.r
\name{user_create}
\alias{user_create}
\title{Create user in the vimo}
\usage{
user_create(user_name, password, mail)
}
\arguments{
\item{user_name}{your user name in **vimo**, character}

\item{password}{your password, character}

\item{mail}{your mail, character}
}
\value{
Information if creating an account was successful.
}
\description{
This function is used to create an user account. You need this to have an access to all features in vimo.
}
\examples{
\\code{
library("vimo")

create_user("Example user", "example password", "example_mail@gmail.com")
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
