% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.r
\name{create_user}
\alias{create_user}
\title{Get an info about the model in **vimo**}
\usage{
create_user(user_name, password, mail)
}
\arguments{
\item{user_name}{your user name in **vimo**, character}

\item{password}{your password, character}

\item{mail}{your password, character}
}
\value{
named list containing all meta data about model
}
\description{
Thanks to this function you can create user on **vimo**.
}
\examples{
library("vimo")

model_info("example_model")

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
