% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_model.r
\name{search_model}
\alias{search_model}
\title{Find model that interests you the most}
\usage{
search_model(tags)
}
\arguments{
\item{tags}{vector of tags, should be strings}
}
\value{
vector of models' names that have at least one common tag with those passed as the argument
}
\description{
Function finds all models having at least one common tag with those passed as the argument
}
\examples{
\\code{
library("vimo")

search_model(c('example', 'easy'))
}

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
