% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{upload}
\alias{upload}
\title{Upload model to **vimo**}
\usage{
upload(model, model_name, model_desc, tags, train_dataset,
  train_dataset_name, dataset_desc, user_name, password)
}
\arguments{
\item{model}{mlr learner, or path to the mlr model written as learner}

\item{model_name}{name of the model that will be visible in **vimo**}
}
\description{
This tool is allows you to upload your *mlr* model to model governance base **vimo**.
}
\examples{
library("vimo")
library("mlr")

task = makeClassifTask(data = iris, target = 'Species')
model = makeLearner("classif.randomForest")
model = train(model, task)

upload(model, "example_model", iris, "example_training_data")

}
\references{
\href{http://192.168.137.64/models}{\bold{models}}
\href{http://192.168.137.64/datasets}{\bold{datasets}}
}
